feat: extract settings into config
